requirements.txt
amqp==5.1.1
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
billiard==4.1.0
celery==5.3.4
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
Jinja2==3.1.2
kombu==5.3.2
MarkupSafe==2.1.3
prompt-toolkit==3.0.39
psycopg2-binary==2.8.4
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
render==1.0.0
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
vine==5.0.0
wcwidth==0.2.9
